feat(Translation): Add Control and Ranked Leagues
From wiki, Ranked Leagues is a game mode introduced in Season 2. It differs from the normal Play Apex game mode in that players are matched with other players of the same or adjacent tier.

The Ranked Leagues seasons coincide with the main seasons. At the end of a season, players receive a reward based on their rank.

I also found something called "APEX ELITE" which is gamemode that ran on June to July of 2019.